requirements.txt
faiss-cpu
Flask==2.2.2
glob2==0.7
matplotlib==3.7.1
numpy==1.23.5
opencv-python-headless
pandas==2.0.1
scikit_image==0.19.3
scikit_learn==1.2.1
tensorflow
gradio
tqdm==4.64.1
